refactor(Application): Dto를 통한 OrderTables생성 로직 위치 변경
- TableGroupService -> TableGroupValidator로 변경
