+++ src/main/java/kitchenpos/application/tablegroup/TableGroupValidator.java
@@ -1,11 +1,17 @@
 package kitchenpos.application.tablegroup;
 
+import java.util.List;
+import java.util.stream.Collectors;
+
 import org.springframework.stereotype.Component;
 
 import kitchenpos.application.order.OrderService;
 import kitchenpos.application.table.TableService;
 import kitchenpos.domain.table.OrderTables;
+import kitchenpos.dto.table.OrderTableDto;
+import kitchenpos.dto.tablegroup.TableGroupDto;
 import kitchenpos.exception.order.HasNotCompletionOrderException;
+import kitchenpos.exception.table.NotRegistedMenuOrderTableException;
 
 @Component
 public class TableGroupValidator {
@@ -32,5 +38,23 @@ public class TableGroupValidator {
         this.validateForUnGroup(orderTables);
 
         return orderTables;
+    }
+
+    public OrderTables getValidatedOrderTables(TableGroupDto tableGroup) {
+        final List<Long> orderTableIds = tableGroup.getOrderTables().stream()
+                                                    .map(OrderTableDto::getId)
+                                                    .collect(Collectors.toList());
+
+        final OrderTables savedOrderTables = OrderTables.of(tableService.findAllByIdIn(orderTableIds));
+
+        checkAllExistOfOrderTables(tableGroup.getOrderTables(), savedOrderTables);
+
+        return savedOrderTables;
+    }
+    
+    private void checkAllExistOfOrderTables(final List<OrderTableDto> orderTables, final OrderTables savedOrderTables) {
+        if (orderTables.size() != savedOrderTables.size()) {
+            throw new NotRegistedMenuOrderTableException();
+        }
     }
 }
